+++ src/umd-index.ts
@@ -1,0 +1,3 @@
+import { LeoWebMonitor } from './core/leo-web-monitor';
+
+export default LeoWebMonitor;